fix(rust): rename l10::l10n to l10n::model
