+++ rust/agama-server/src/l10n.rs
@@ -21,16 +21,10 @@
 mod dbus;
 pub mod error;
 pub mod helpers;
-mod keyboard;
-pub mod l10n;
-mod locale;
-mod timezone;
+mod model;
 pub mod web;
 
 pub use agama_lib::localization::model::LocaleConfig;
 pub use dbus::export_dbus_objects;
 pub use error::LocaleError;
-pub use keyboard::Keymap;
-pub use l10n::L10n;
-pub use locale::LocaleEntry;
-pub use timezone::TimezoneEntry;
+pub use model::{Keymap, L10n, LocaleEntry, TimezoneEntry};
