+++ rust/agama-server/src/l10n/model/locale.rs
